data transformer and dataset changes
In this commit there has been an attempt made to data transformers and data augmentors mainly in the hope to keep it as original
as possible there are still some discrepancy which should be resolved in the future currently only aug transform none works 100%
as expected. Main areas to look out for is cropping and places where its used and possibly change it to dynamic cropping in
future.

On this end, two new functions are added to data_augmentors but this needs full testing, currently not in use.