add 3 new crop methods
Method 0:
Original crop from deep-prio++ for msra dataset WITH OFFSET ADDITION REMOVED (note this is imp!) as this is not needed for FHAD and
produces undesirable results, this has no impact on msra provided only meth 0, 1, 3 are used i.e. no custom aspect ratio cropping.

Method 1:
  We supply keypt values centered w.r.t to CoM i.e keypt_diff = keypt - CoM then take the max(abs(keypt_diff)) of each dir and add some
extra mm padding in each direc (+ and - sperately) with equal amounts of padding for x,y with max_val(x_max, y_max) chosen.
  similar for z using redefined padding value.

Method 2:
  Here all stuff is done in pixel domain, find first the min and max of, x,y values and then for each dir choose whichever is farthest
from com and set that as 'amount' of buffer for xy_startend
  zstart is based on depth_pad param and max and min keypts value so not really linked to com....

Method 3:
 An improvement of method 2 based on com now, a px_diff is now computed much similar to method 1. the x and y diffs are compared and
max of that is selected for both x and y for aspect ratio preserving and for zaxis as well com is used so that in end com is always in
center of point for x,y and z axis. the buffer is added respectively for x, y and z for x,y, exact same values are highly recommended
for aspect ratio preservation.

Although visually Method #2 is bad, suprisingly, error-wise it gives the best results. We compared on real-test set data and found this
performed the best. For reference this was BaselineHPE/0419_1744. Ones that are worse include BaselineHPE/0418_1050,
BaselineHPE/0418_0205 and BaselineHPE/0418_0011. We need to affirm whether we can still get back good visual results by using this
method, train and then after inference using another crop method to get predicted values to 2D image.