added new data loading procedure and deterministic aug for FHAD dataset
for data loading procedure we now support -1.0 <= val_split <= 0 as a fraction of the 'test-set' rather than the trian-set. We will
probably use val-split = -0.2 in future as it closely matches full test set performance. We did this because PCA was being used on
entire train-set and also it wasn't a good tracker for the test-set accuracy i.e. val_split 0.2 was very off from val_Set as full test
set. This -0.2 is mainly for faaster validation but also to ensure we don't fully overfit to val-set.