attention: it works!
finally after experimenting with many different architecture a working attention arch

this gives better results (slightly) in lstm to about 0.725 from 0.723 in lstm gt keypt performance and when used
in conjuction for fine tuning in combined model with action feedback initally its much worse like 0.29 to begin with
and then it gets a lot better as we train so the fine tuning helps and soon about 15ep it overtakes in val acc.

for lstm this is v5 for combined model this is 6d model.

we have also included a new model which is averaging of y using moving average which provides temporal smoothing mechanism this is
untested at the moment only delivers train improvements but not test this is v8d4.