backup commit for hpe changes
in baseline hpe model we have added several ways to train with noisy conditioning data, we test with probabilistic conditioning but
it makes the models worse after 15 epochs, if we decrease the probability it needs to be done by a large margin and thus the error
does reduce but when we perform testing its still not robust to noise, basically the only way for noise robustness was to create a
model that performs as close as possible preferrably a lower bound to non conditioned hpe, then later we use some noisy feedback of
action to see for any improvements.

only thing that shows improvements during test time over baseline using sequence based noisy data is pretraining with equiprob cond
only the resnet area. we choose only resnet because we dont want strong dependence on equiprob action, otherwise we suspect that
train time imrpovements will be harder to show (need to prove this somehow)

this is ver7.0 of action cond. we've added a few more beyond that equiprob conditioning that doesnt perform well...